Append events instead of overwritting
This is to ensure that events are not deleted.

An issue with updates to the events JSON field from being saved was encountered.
Multiple attempts were made to resolve this issue including the use of
sqlalchemy.ext.mutable and options details https://stackoverflow.com/questions/60893464/nested-json-array-not-updating-in-db/60927486#60927486
however none of these proved successful. As such we simple mark the field
as dirty if additional events are inserted.