CRM457-1905: Check submission state (#257)
## Description of change
We found an issue where submissions were sent to the app store with
status 'draft'. But we didn't spot it immediately because we were
silently 'correcting' the status in the state field, but not in the
"blob". Let's be more picky, and fail out immediately if we spot
impermissible data

[Link to relevant
ticket](https://dsdmoj.atlassian.net/browse/CRM457-1905)