Stop locking submission while notifying (#301)
## Description of change
The E2E tests run sidekiq jobs synchronously. Locking the submission
while notifying subscribers means if the subscriber makes a synchronous
call back to the app store that modifies the submission, we get a
deadlock.

If the notify subscriber job doesn't lock, then we can't call it inside
another lock, otherwise it may run with non-fresh data (if it runs
before the outer lock has been released).

So we move the calls to NotifySubscriber outside locks. This means if
Redis is down the DB will be updated without the job queue being
updated, so the apps will get out of sync with one another. However our
`notify_subscriber_completed` flag will give us a paper trail to fix
this problem if it occurs

[Link to relevant
ticket](https://dsdmoj.atlassian.net/browse/CRM457-2176)