Update search logic to improve clarify and correctness
When the search was first written, having special characters in thequery
wasn't considered other than to attempt to trim some of them; causing
some confusing syntax error exceptions.

Instead we escape any non-numeric words that we can and allow a wider
amount of queries to be supported, as well as adjusting the search
vector fields to support partial matching of defendent names.